+++ src/components/common/results/Divider.jsx
